ref: Move FilterState into FilterView
This means the DockState owns the FilterStates and is thus moved into the LogView. This entails a new sub-struct LogViewState and many small changes.

Alternatively, we could have passed the DockState into logview, when needed, but it anyway belongs in there.

Saving from within the filter is not possible anymore as they cannot access all tabs (circular ownership).